增加了15日天气趋势界面，自定义plotView用于显示15日天气变化的折线图

+++ app/src/main/java/com/example/liziweather/logic/model/RealtimeWeatherResponse.kt
@@ -56,7 +56,7 @@ data class RealtimeWeatherResponse(
     )
 
     data class Wind(
-        val direction: Int,
+        val direction: Double,
         val speed: Double
     )
